refactor: Update import paths for currency module
