+++ Commands/CurrencyCommands/admin_add_currency_command.py
@@ -1,9 +1,9 @@
 from discord.ext import commands
 from discord import app_commands
 import discord
-from data.currency import add_bananas
+from data.Currency.currency import add_bananas
 from utils.users import UNBUTTERED_BAGEL_ID
-
+from data.nickname import get_nickname
 async def define_admin_add_currency_command(tree, servers):
     def is_owner():
         async def predicate(interaction: discord.Interaction):
@@ -18,7 +18,7 @@ async def define_admin_add_currency_command(tree, servers):
             return
         try:
             await add_bananas(str(user.id), amount)
-            await interaction.response.send_message(f"Added {amount} bananas to {user.display_name}'s account.")
+            await interaction.response.send_message(f"Added {amount} bananas to {get_nickname(user.id) or user.display_name}'s account.")
         except Exception as e:
             await interaction.response.send_message(f"Failed to add currency: {str(e)}")
