+++ Commands/CurrencyCommands/daily_currency_command.py
@@ -1,4 +1,5 @@
-from data.daily import try_collect_daily
+from data.Currency.daily import try_collect_daily
+from data.nickname import get_nickname
 from discord.ext import commands
 from discord import app_commands
 import discord
@@ -20,6 +21,6 @@ async def define_daily_command(tree, servers):
             bananas_collected = result
             model = "gpt-4o"
             story = f"You are a narrator talking about how somebody came across money. You do this in 1 or 2 lines and always mention the user along with how many banana coins, you can also use the emoji {BANANA_COIN_EMOJI}"
-            user_input = f"{interaction.user.mention} collected {bananas_collected} bananas."
+            user_input = f"{get_nickname(user_id) or interaction.user.mention} collected {bananas_collected} bananas."
             response_message = await generate_gpt_response(model, story, user_input)
             await interaction.followup.send(response_message)
