+++ Commands/CurrencyCommands/get_currency_command.py
@@ -1,7 +1,7 @@
 from discord.ext import commands
 from discord import app_commands
 import discord
-from data.currency import get_bananas
+from data.Currency.currency import get_bananas
 from utils.emoji_helper import BANANA_COIN_EMOJI
 
 async def define_get_currency_command(tree, servers):
