+++ Commands/CurrencyCommands/send_bananas_command.py
@@ -1,7 +1,7 @@
 import discord
 from discord.ext import commands
 from discord import app_commands
-from data.currency import get_bananas, add_bananas, remove_bananas
+from data.Currency.currency import get_bananas, add_bananas, remove_bananas
 from utils.emoji_helper import BANANA_COIN_EMOJI
 class ConfirmView(discord.ui.View):
     def __init__(self, user: discord.User, amount: int, initiator: discord.User, channel: discord.TextChannel):
