+++ Commands/GamblingCommands/coinflip_command.py
@@ -2,7 +2,7 @@ from discord.ext import commands
 from discord import app_commands
 import discord
 import random
-from data.currency import get_bananas, add_bananas, remove_bananas
+from data.Currency.currency import get_bananas, add_bananas, remove_bananas
 from game.shared_logic import bet_checks
 from utils.emoji_helper import BANANA_COIN_EMOJI
 async def define_coinflip_command(tree, servers):
