+++ Commands/ShopCommands/assassinate_command.py
@@ -2,7 +2,7 @@ from datetime import timedelta
 from discord.ext import commands
 from discord import app_commands
 import discord
-from data.currency import get_bananas, remove_bananas
+from data.Currency.currency import get_bananas, remove_bananas
 from utils.emoji_helper import BANANA_COIN_EMOJI
 
 async def define_assassinate_command(tree, servers):
