+++ Commands/ShopCommands/shop_command.py
@@ -1,7 +1,7 @@
 import discord
 from discord.ui import View, Button
 from data.items import add_item
-from data.currency import get_bananas, remove_bananas
+from data.Currency.currency import get_bananas, remove_bananas
 from utils.emoji_helper import BANANA_COIN_EMOJI
 
 class ShopView(View):
