+++ Commands/StatCommands/reroll_luck_command.py
@@ -2,7 +2,7 @@ import discord
 from discord import app_commands
 from discord.ui import View, Button
 from data.items import add_item
-from data.currency import get_bananas, remove_bananas
+from data.Currency.currency import get_bananas, remove_bananas
 from data.stats import random_luck, set_luck
 from utils.emoji_helper import BANANA_COIN_EMOJI
